# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Biblioteca",
        "description": "Documentação da API de gerenciamento de biblioteca",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/V1/loans": {
            "get": {
                "tags": [
                    "Loans"
                ],
                "summary": "Lista todos os empréstimos",
                "operationId": "241dc5105b5ca3e6111f4a4ef4529a1a",
                "responses": {
                    "200": {
                        "description": "Lista de empréstimos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Loans"
                ],
                "summary": "Cria um novo empréstimo",
                "operationId": "c000e0c1f6b528505d5e87e2d1470e62",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "book_id",
                                    "loan_date",
                                    "return_date",
                                    "status"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "string"
                                    },
                                    "book_id": {
                                        "type": "string"
                                    },
                                    "loan_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "return_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "returned",
                                            "overdue"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Empréstimo criado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/V1/loans/{id}": {
            "put": {
                "tags": [
                    "Loans"
                ],
                "summary": "Atualiza um empréstimo existente",
                "operationId": "76cda1f8ddb4c66d4cc5881b5bda83b4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "book_id",
                                    "loan_date",
                                    "return_date",
                                    "status"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "string"
                                    },
                                    "book_id": {
                                        "type": "string"
                                    },
                                    "loan_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "return_date": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "returned",
                                            "overdue"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empréstimo atualizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Loans"
                ],
                "summary": "Exclui um empréstimo",
                "operationId": "c573e72dc7299876f0d1e2fc42b3e5ce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empréstimo excluído"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}